requirements.txt
rich>=13.7.0
typer>=0.9.0

open-clip-torch>=2.24.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.2.0

python-magic>=0.4.27
pypdf>=4.0.0
python-docx>=1.1.0

tqdm>=4.66.0
